Read correctly index documents defined with floats.
That's what generally ends up in them, whenever the shell is used
to define the index. Bug reported by Nick Stott.

Also fix the index names to include the kind. It uses to not include
the name because the shell had a bug, and mgo preserved the same bug,
but now that the shell is fixed there's no reason to keep the broken
behavior.