txn: fix transactions with multiple ops on same document
The application of multiple operations on the same document id in the
same transaction was not functioning properly. All but the first
transaction would fail to apply due to miscalculation of the sequence
of revnos.

So, for example, if a a transaction was doing Insert+Update on the
same collection and id, the Update was never applied, whether the
document existed or not. If the transaction was doing Update+Insert,
the Update is currently applied if the document exists, and
the Insert is never applied.

This change fixes that. Insert+Update will now work properly if the
document exists or not (if it exists only the Update will do something).
Doing Update+Insert is a nice to way to say "either update or insert",
as only a single one of them can possibly work depending on whether
the document exists or not.

This change should not affect existent code, unless the code was
already buggy and not functioning properly.