Introduced FailFast option in DialInfo.
FailFast will cause connection and query attempts to fail faster when
the server is unavailable, instead of retrying until the configured
timeout period. Note that an unavailable server may silently drop
packets instead of rejecting them, in which case it's impossible to
distinguish it from a slow server, so the timeout stays relevant.

Feature requested by Louisa Berger and Cailin Nelson.