Drop support for custom index name on EnsureIndex.
MongoDB only supports a single index with a given index key, so
the custom name is a distraction. It will create surprising behavior
where a given index is not created because a different index with
the same key already exists, and it breaks the use of the current
API, such as with DropIndex which is based on the index key.

Updates #65.