Avoid MongoDB pipeline size limit in PurgeMissing
Under MongoDB 2.4 and earlier, aggregation pipeline results may be no
bigger than 16MB, even when an iterator is used. It was possible for
PurgeMissing to hit this limit when dealing with large txn-queue
arrays, causing it to fail.

This change refactors PurgeMissing so that it no longer uses an
aggregation pipeline, working around the limitation. A regression test
is included which triggered the failure mode for the previous version
of PurgeMissing.

The aggregation pipeline result size limitation was removed in MongoDB
2.6. See: http://docs.mongodb.org/manual/core/aggregation-pipeline-limits/