Replace bad not found check in Update.
`Update` was assuming that, if no document was updated, the selector query wasn't matching anything, thus returning `ErrNotFound`. That's wrong, because it can happen that the selector matches but then the new updated value is jus the same as the old one, in which case `nModified` is 0 even if the query is successful.

Now `LastError` has a `Found` method that returns true if `result.N` is greater than zero. `Update` checks that method instead of `UpdatedExisting`.

I haven't added a test nor run the existing ones because apparently they need a quite complex previous setup in your machine. Sorry for that.