txn: preserve Op fields when loading transactions
When loading transactions, previously the code would
unmarshal directly into Op structs. This meant that
Assert, Insert and Update values would take on the
default representations, since those fields in Op
have type interface{}.

If, for example, you use a struct in an assertion,
then it would be marshalled as a map, with the correct
order. Upon unmarshalling into interface{}, it would
become a bson.M and the order would be lost.

We introduce Op.SetBSON to first unmarshal into
bson.Raws, and then set those on the Assert, Insert
or Update fields. This preserves the values exactly.