Reduce excessive timeout delay on initial dial.
The pinger was being run synchronously to obtain an initial
ping value on mongoServer creation, which meant a large delay
when the server was not available. Now the server is initialized
with a large ping value to penalize it until the background
pinger has a chance to update the value, and the server is
returned immediately.

Reported by John Morales on https://jira.mongodb.org/browse/MGO-54

Fixes #142.