bson: add AsUpdate
Add an AsUpdate function to determine what fields will be
updated by a given document.

We need to to add a document into mongodb, atomically incrementing a
version counter if the document already exists.

The obvious answer is to use Upsert, but to do what we want there, we'd
need to explicitly mention every field in the document (it has quite a
few fields and we add more relatively often).

Say our document type looked like this:

        type mdoc struct {
                Id string       `bson:"_id"`
                Field1 int
                Field2 string
                Field3 time.Time       `bson:",omitempty"`
                Field4 string       `bson:",omitempty"`
        }

then our code for upserting the document would need to look
something like this:

        func addDoc(id string, doc mdoc) {
                set := bson.M{
                        "field1": doc.Field1,
                        "field2": doc.Field2,
                }
                unset := make(bson.M)
                if doc.Field3.IsZero() {
                        unset["field3"] = nil
                } else {
                        set["field3"] = doc.Field3
                }
                if doc.Field4 == "" {
                        unset["field4"] = nil
                } else {
                        set["field4"] = doc.Field4
                }
                c.UpsertId(id, bson.M{
                        "$set": set,
                        "$unset": unset,
                        "$inc", bson.M{"version", 1},
                })
        }

We'd much prefer to have the fields automatically derived from
the document type, so when we add a new field, we don't have
to fix the logic and there's less chance of the fields getting
out of step with the type.

There's no great way to do that AFAICS; hence our suggestion
for AsUpdate. With that, we could write the logic as:

        func addDoc(id string, doc mdoc) {
                u, err := bson.AsUpdate(doc)
                if err != nil {...}
                delete(u.Set, "version")
                c.UpsertId(id, bson.M{
                        "$set": u.Set,
                        "$unset": u.Unset,
                        "$inc", bson.M{"version", 1},
                })
        }

We think that it's a good fit for the bson package itself because
the code to implement AsUpdate depends on the specific
marshaling logic within the bson package, which would
need to be copied if we were to implement the functionality
independently.