txn: return error from flusher.checkpoint
If the call to f.tc.Update in flusher.checkpoint
fails due to an error other than mgo.ErrNotFound,
the error is currently swallowed. flusher.advance
will then loop back around and call checkpoint
again, which may continue to fail (e.g. if the
Mongo connection is in a bad state), pegging the
CPU.