use active sock count to control limit on pool (#6)
Due to a concurrency bug mgo driver is not always put upper limit on max connections in a pool. This bug can cause more connections to open to mongo.

In this code change we're using socket state and a placeholder to keep our slot in aciveSockets slice. This will ensure that we dont run over pool limit.